Заготовка

+++ ReportEngine.Shared/Config/IniHeleprs/CalculationSettings/HumanCostSettings.cs
@@ -4,15 +4,15 @@ public class HumanCostSettings
 {
     public float ObvzyakaProduction { get; set; } // Изготовление обвязок
     public float CollectorProduction { get; set; } // Изготовление коллектора
-    public double Tests { get; set; } // Испытания
-    public double CommonCheckStand { get; set; } // Общая проверка стенда
-    public double TimeForCheckStand { get; set; } // Время на проверку 1 стенда
-    public double TimeForFinalWork { get; set; }  // Время на финальную работу
-    public double TimeForOneDrill { get; set; } // Время на сверление 1 отверстия
-    public double TimeForCollectorBoil { get; set; } // Время варки коллектора
-    public double TimeForAllChecks { get; set; } // Время проведения всех испытаний
-    public double TimeForPrepareAllEquipment { get; set; } // Время подготовки всего оборудования
-    public double TimeForDrillOneBus { get; set; } // Время на сверления одной шины 
-    public double TimeForMontageOneInput { get; set; } // Время монтажа одного ввода
-    public double TimeForOthersOperations { get; set; }  // Время на другие операции
+    public float Tests { get; set; } // Испытания
+    public float CommonCheckStand { get; set; } // Общая проверка стенда
+    public float TimeForCheckStand { get; set; } // Время на проверку 1 стенда
+    public float TimeForFinalWork { get; set; }  // Время на финальную работу
+    public float TimeForOneDrill { get; set; } // Время на сверление 1 отверстия
+    public float TimeForCollectorBoil { get; set; } // Время варки коллектора
+    public float TimeForAllChecks { get; set; } // Время проведения всех испытаний
+    public float TimeForPrepareAllEquipment { get; set; } // Время подготовки всего оборудования
+    public float TimeForDrillOneBus { get; set; } // Время на сверления одной шины 
+    public float TimeForMontageOneInput { get; set; } // Время монтажа одного ввода
+    public float TimeForOthersOperations { get; set; }  // Время на другие операции
 }
